% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/stdea.R
\name{stdea}
\alias{stdea}
\title{Spatio-Temporal DEA efficiency}
\usage{
stdea(X, Y, RTS = "vrs", ORIENTATION = "out")
}
\arguments{
\item{X}{Inputs of firms to be evaluated, a K x m matrix of observations of K firms with m inputs (firm x input). In case TRANSPOSE=TRUE the input matrix is transposed to input x firm.}

\item{Y}{Outputs of firms to be evaluated, a K x n matrix of observations of K firms with n outputs (firm x input). In case TRANSPOSE=TRUE the output matrix is transposed to output x firm.}

\item{RTS}{Text string or a number defining the underlying DEA technology / returns to scale assumption.}

\item{ORIENTATION}{Input efficiency "in" (1), output efficiency "out" (2), and graph efficiency "graph" (3). For use with DIRECT , an additional option is "in-out" (0).}
}
\value{
The results are returned in a Farrell object with the following components. The last three components in the list are only part of the object when SLACK=TRUE.
}
\description{
Estimates a ST-DEA frontier and calculates efficiency measures a la Farrell.
}

